# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Food Court Users API
  version: 1.0.0
  description: Gestión de usuarios de la plazoleta de comidas.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      operationId: login
      summary: Iniciar sesión
      description: Autentica mediante correo y contraseña.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Autenticación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseEnvelope'
        '400':
          description: Formato de solicitud inválido
        '401':
          description: Correo o contraseña incorrectos
  /users/owners:
    post:
      operationId: createOwner
      summary: Crear un propietario
      description: Crea un usuario propietario..
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerRequest'
      responses:
        '201':
          description: Propietario creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponseEnvelope' }
        '400': { description: Datos inválidos, propietario menor de edad, email o documento duplicado, content: { application/json: { schema: { $ref: '#/components/schemas/ExceptionResponse' } } } }
        '401':
          description: Token ausente o inválido
        '403':
          description: Se requiere rol ADMIN
  /users/employees:
    post:
      operationId: createEmployee
      summary: Crear un empleado
      description: Crea un empleado y lo asocia a un restaurante del propietario autenticado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeRequest'
      responses:
        '201':
          description: Empleado creado y asociado al restaurante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseEnvelope'
        '400':
          description: Datos inválidos, correo o documento duplicado
        '401':
          description: Token ausente o inválido
        '403':
          description: El usuario no es propietario del restaurante
        '404':
          description: El restaurante no existe
        '503':
          description: El microservicio de plazoleta no está disponible
  /users/{userId}/role:
    get:
      operationId: getUserRole
      summary: Consultar el rol de un usuario
      description: Obtener rol de un usuario, usado por el microservicio de plazoleta para validar propietarios.
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        '200':
          description: Rol encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserRoleResponseEnvelope' }
        '404':
          description: Usuario inexistente
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExceptionResponse' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          writeOnly: true
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        tokenType:
          type: string
          example: Bearer
    LoginResponseEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/LoginResponse'
    OwnerRequest:
      type: object
      required:
        - name
        - lastName
        - identityDocument
        - cellphone
        - birthDate
        - email
        - password
      properties:
        name: { type: string, minLength: 1 }
        lastName: { type: string, minLength: 1 }
        identityDocument: { type: string, pattern: '^\d+$' }
        cellphone: { type: string, maxLength: 13, pattern: '^\+?\d+$', example: '+573005698325' }
        birthDate: { type: string, format: date }
        email: { type: string, format: email }
        password: { type: string, format: password, writeOnly: true }
    EmployeeRequest:
      type: object
      required:
        - name
        - lastName
        - identityDocument
        - cellphone
        - email
        - password
        - roleId
        - restaurantId
      properties:
        name:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        identityDocument:
          type: string
          pattern: '^\d+$'
        cellphone:
          type: string
          maxLength: 13
          pattern: '^\+?\d+$'
          example: '+573005698325'
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          writeOnly: true
        roleId:
          type: integer
          format: int64
          example: 3
        restaurantId:
          type: integer
          format: int64
          minimum: 1
    UserResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
        roleId: { type: integer, format: int64 }
        role: { type: string, example: OWNER }
    UserResponseEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/UserResponse' }
    UserRoleResponse:
      type: object
      required: [id, role]
      properties:
        id: { type: integer, format: int64 }
        role: { type: string, enum: [ADMIN, OWNER, EMPLOYEE, CUSTOMER] }
    UserRoleResponseEnvelope:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/UserRoleResponse' }
    ExceptionResponse:
      type: object
      properties:
        errors:
          type: object
          additionalProperties: { type: string }